feat: implement find all book

+++ src/main/java/RickaPrincy/model/Book.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 public class Book {
     private String id, bookName;
-    private int pageNumbers;
+    private long pageNumbers;
     private Date releaseDate;
     private List<Topic> topics;
     private Author author;
